Fixed not waiting for complete writing of bin stacks file

+++ src/dotnet/Procfiler/src/Commands/CollectClrEvents/ICommandExecutorDependantOnContext.cs
@@ -128,8 +128,13 @@ public class CommandExecutorImpl(
       }
       finally
       {
-        process.Kill();
-        process.WaitForExit();
+        const int WaitForProcessTimeoutMs = 10_000;
+        if (!process.WaitForExit(WaitForProcessTimeoutMs))
+        {
+          logger.LogWarning("Failed to wait ({Timeout}ms) until process terminates naturally, killing it", WaitForProcessTimeoutMs);
+          process.Kill();
+          process.WaitForExit();
+        }
       }
 
       if (context.CommonContext.PrintProcessOutput)
